refactor: Use `std::uint32_t` to index new space points and seeds (#4394)
32 bit seems sufficient. no need to go up to 64 bits

+++ Core/include/Acts/EventData/Types.hpp
@@ -15,10 +15,10 @@
 
 namespace Acts {
 
-using SpacePointIndex2 = std::size_t;
+using SpacePointIndex2 = std::uint32_t;
 using SpacePointIndexRange2 = std::pair<SpacePointIndex2, SpacePointIndex2>;
 
-using SeedIndex2 = std::size_t;
+using SeedIndex2 = std::uint32_t;
 
 using TrackIndexType = std::uint32_t;
 static constexpr TrackIndexType kTrackIndexInvalid =
